+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### 0.6.0 Unreleased
 
+- Raise when trying to cache a belong_to association with a scope. Previously the scope was ignored on a cache hit (#323)
 - Remove deprecated `never_set_inverse_association` option (#319)
 
 #### 0.5.1